Move CommandChannel into the PacketTunnelActor namespace for consistency with Command
